Add anytime and incremental lattice computation.
The computation of bigger clusters time out and we lose any mining results. This change enable us to:
- do the binning and lattice construction incrementally (i.e., after processing a single groum)
- save the intermediate result of the mining.

This way we do not need to process all the graphs in a bin to get a usable result.

The process can be also re-started at any point in time to improve the results.